Moved copilot component to claimReviewEditor file
